feat: New storage interface moved to separate module
