+++ component/newstorage/edv/restclient.go
@@ -15,8 +15,8 @@ import (
 	"net/http"
 	"net/url"
 
+	"github.com/hyperledger/aries-framework-go/component/newstorage"
 	"github.com/hyperledger/aries-framework-go/pkg/common/log"
-	"github.com/hyperledger/aries-framework-go/pkg/newstorage"
 )
 
 var logger = log.New("EDV-Provider")
