+++ component/newstorage/go.mod
@@ -1,0 +1,5 @@
+// Copyright SecureKey Technologies Inc. All Rights Reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+module "github.com/hyperledger/aries-framework-go/component/newstorage"
