+++ component/newstorage/newstorage.go
