+++ test/newstorage/go.mod
@@ -8,8 +8,9 @@ go 1.15
 
 require (
 	github.com/google/uuid v1.1.2
-	github.com/hyperledger/aries-framework-go v0.0.0
+	github.com/hyperledger/aries-framework-go/component/newstorage v0.0.0
 	github.com/stretchr/testify v1.6.1
 )
 
-replace github.com/hyperledger/aries-framework-go => ../..
+replace github.com/hyperledger/aries-framework-go/component/newstorage => ../../component/newstorage
+
